test: update tests for refactorings
